requirement.txt
opencv-contrib-python          
opencv-contrib-python-headless 
opencv-python                  
opencv-python-headless  
requests
pillow       
torchvision
